Add: pip3を探す処理を追加した

+++ bin/build-vim.sh
@@ -13,13 +13,23 @@ install_vim_linux_yum() {
   sudo yum groups install -y 'Development tools'
   sudo yum install -y cmake
   if type -p pip3 2>/dev/null; then
-    sudo pip3 install neovim --upgrade
+    pip3=pip3
   else
-    py3_ver=$(yum list python3\*-pip -q | grep -Eo '^python3[0-9]+' | sort | tail -n1)
-    sudo yum install -y "${py3_ver}"-{pip,devel}
-    pip3=$(rpm -ql "${py3_ver}-pip" | grep -Eo '^([/a-z]*)/bin/pip.*3.*' | head -n1)
-    sudo "$pip3" install neovim --upgrade
+    py3_ver=$(yum list python3\*-devel -q | grep -Eo '^python3[0-9]+' | sort | tail -n1)
+    [[ -z $python3 ]] && return 1
+    sudo yum install -y "${py3_ver}"-devel
+    if yum list "${py3_ver}-pip" >/dev/null 2>&1; then
+      sudo yum install -y "${py3_ver}"-pip
+      pip3=$(rpm -ql "${py3_ver}-pip" | grep -Eo '^([/a-z]*)/bin/pip.*3.*' | head -n1)
+    else
+      curl https://bootstrap.pypa.io/get-pip.py | sudo python3
+      if type -p pip3 2>/dev/null; then
+        pip3=pip3
+      fi
+    fi
   fi
+  [[ -z $pip3 ]] && return 1
+  sudo "$pip3" install neovim --upgrade
   (
     cd "$(mktemp -d)" || exit 1
     trap "rm -rf $(printf %q "$(pwd)")" EXIT
